+++ engine/Poseidon/Core/Game/GameLoop.cpp
@@ -209,7 +209,7 @@ bool AppIdle()
     lastTime = actTime;
 
     static DWORD lastSysTime;
-    DWORD sysTime = ::GetTickCount();
+    DWORD sysTime = static_cast<DWORD>(::GetTickCount());
     DWORD timeDelta = sysTime - lastSysTime;
     lastSysTime = sysTime;
